Don't treat KeyboardInterrupt specially in pager context
If a KeyboardInterrupt occurs while the pager context manager is active,
don't treat it different to other kinds of exceptions. We still need to
ignore Ctrl-C while we're waiting for the pager to exit, but whatever
exception (if any) caused the context manager to exit should be bubbled
up to the caller and handled at a higher level.