Don't support input directly from the terminal
If input is from stdin but there is no pipe or file redirecting to it,
that's almost certainly due to forgetting to put the file name on the
command line rather than an intention to type the logs into the
terminal. In this case, just exit immediately rather than sit and wait
for input with the pager active (which is confusing until you realise
the mistake).