Handle errors from controller-runtime
We now get reconciler errors from the logger
"controller-runtime.manager.controller.baremetalhost", which is in the
RUNTIME group rather than the CONTROLLER group. This log has "name" and
"namespace" keys and should be treated like other CONTROLLER logs for
the purposes of getting the name. Only PROVISIONER logs have the
"namespace~name" format, so restrict looking for it to the provisioner
logs.